refactor(app): share contracts and isolate builds
