+++ packages/vera-app/src/renderer/types.ts
@@ -1,3 +1,29 @@
+import type {
+  AppSettings,
+  ChatCitationResult,
+  CredentialResult,
+  Session,
+  StreamEvent,
+} from '../shared/contracts';
+
+export type {
+  AppSettings,
+  ChatAttachment,
+  ChatCitationResult,
+  ContentPart,
+  ContextChunkResult,
+  CredentialResult,
+  FigureResult,
+  ProviderProfile,
+  RegionResult,
+  SearchResult,
+  Session,
+  SessionTurn,
+  StreamEvent,
+  TraceMessage,
+  TraceToolCall,
+} from '../shared/contracts';
+
 export interface VeraResponse<T = unknown> {
   id?: string;
   ok: boolean;
@@ -5,48 +31,6 @@ export interface VeraResponse<T = unknown> {
   error?: string;
   traceback?: string;
   cancelled?: boolean;
-}
-
-export type ContentPart =
-  | { type: 'text'; text: string }
-  | { type: 'image_url'; image_url: { url: string } };
-
-export interface TraceMessage {
-  role: string;
-  content?: string | ContentPart[] | null;
-  name?: string;
-  tool_call_id?: string;
-  tool_calls?: { id?: string; type?: string; function?: { name?: string; arguments?: string } }[];
-}
-
-export interface TraceToolCall {
-  id?: string;
-  name?: string;
-  arguments?: Record<string, unknown>;
-}
-
-export interface StreamEvent {
-  id: string;
-  event: 'search_start' | 'search_done' | 'llm_request' | 'llm_response' | 'tool_call' | 'answer_delta' | 'answer_reset';
-  turn?: number;
-  query?: string;
-  mode?: string;
-  top_k?: number;
-  hits?: number;
-  // llm_request
-  model?: string;
-  tools?: string[];
-  messages?: TraceMessage[];
-  // llm_response
-  content?: string;
-  tool_calls?: TraceToolCall[];
-  usage?: Record<string, unknown> | null;
-  // tool_call
-  name?: string;
-  arguments?: Record<string, unknown>;
-  output?: unknown;
-  // answer_delta
-  text?: string;
 }
 
 export interface VeraApi {
@@ -157,62 +141,6 @@ export interface ValidateResult {
   issues: string[];
 }
 
-export interface FigureResult {
-  page_number: number;
-  bbox?: number[];
-  page_width?: number;
-  page_height?: number;
-  asset_id?: string;
-  filename?: string;
-  caption?: string | null;
-  data_url?: string;
-  included_in_context?: boolean;
-}
-
-export interface RegionResult {
-  page_number?: number;
-  bbox?: number[];
-  page_width?: number;
-  page_height?: number;
-}
-
-export interface ContextChunkResult {
-  chunk_id: string;
-  text: string;
-  page_start: number | null;
-  page_end: number | null;
-  heading_path: string | null;
-}
-
-export interface SearchResult {
-  chunk_id: string;
-  score: number;
-  text: string;
-  page_start: number | null;
-  page_end: number | null;
-  heading_path: string | null;
-  source_filename: string | null;
-  document_id: string;
-  regions?: RegionResult[];
-  figures?: FigureResult[];
-  before_chunks?: ContextChunkResult[];
-  after_chunks?: ContextChunkResult[];
-  file?: string;
-}
-
-export interface ChatCitationResult {
-  id: string;
-  label: string;
-  result: SearchResult;
-}
-
-export interface ChatAttachment {
-  id: string;
-  name: string;
-  mime_type: string;
-  data_url: string;
-}
-
 export interface ChatAnswerResult {
   prompt: string;
   answer: string;
@@ -233,30 +161,6 @@ export interface ChatAnswerResult {
   };
 }
 
-export interface SessionTurn {
-  role: 'user' | 'assistant';
-  content: string;
-  citations?: ChatCitationResult[];
-  attachments?: ChatAttachment[];
-  searches?: { query: string; mode: string; top_k: number; hits: number }[];
-  answer_mode?: 'retrieval' | 'agent';
-  mode_label?: string;
-  trace?: StreamEvent[];
-  images_sent?: number;
-  vision_fallback?: boolean;
-  llm?: { provider: string; model: string; usage?: Record<string, unknown> | null };
-  timestamp: number;
-}
-
-export interface Session {
-  id: string;
-  title: string;
-  source_path: string;
-  turns: SessionTurn[];
-  created_at: number;
-  updated_at: number;
-}
-
 export interface Mode {
   id: string;
   label: string;
@@ -271,38 +175,6 @@ export interface Mode {
   max_figure_images: number;
   builtin: boolean;
   path: string;
-}
-
-export interface ProviderProfile {
-  id: string;
-  preset_key?: string;
-  label: string;
-  provider: string;
-  base_url: string;
-  api_key_env: string;
-  auth_type: string;
-  temperature: number;
-  models: string[];
-  available_models?: string[];
-  models_refreshed_at?: number;
-  model_options?: Record<string, {
-    reasoning_effort?: string;
-    fast?: boolean;
-  }>;
-  has_api_key?: boolean;
-}
-
-export interface AppSettings {
-  providers: ProviderProfile[];
-  active_provider_id: string;
-  active_model: string;
-  active_mode_id: string;
-}
-
-export interface CredentialResult {
-  ok: boolean;
-  has_api_key: boolean;
-  error?: string;
 }
 
 export interface ConvertResult {
