+++ packages/vera-app/tsconfig.electron.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "lib": ["ES2022"],
+    "noEmit": false,
+    "outDir": "dist-electron",
+    "types": ["node"]
+  },
+  "include": ["electron/**/*.ts", "src/shared/**/*.ts"],
+  "exclude": ["electron/**/*.test.ts"]
+}
